python-examples/chatgpt-clone-simple-memory-langchain/pyproject.toml
[tool.poetry]
name = "chatgpt-clone-simple-memory-langchain"
version = "0.1.0"
description = ""
authors = ["Aurélien Brabant <aurelien@agentlabs.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.311"
openai = "^0.28.1"
python-dotenv = "^1.0.0"
agentlabs-sdk = "^0.0.33"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
